lxc: fix meson.build on esoteric host systems
When building lxc's meson.build tries to infer
something for the target system out of the host
build OS. This isn't reproducible and can actually
fail on some OS' like NixOS.

The failure looked like this, early in the
building stage of lxc:

> ../../../../build_dir/target-aarch64_cortex-a53_musl/lxc-6.0.5/meson.build:166:8: ERROR: Problem encountered: "distrosysconfdir" is not set

The /etc/default seems to be something that is
derived on most host systems, so use that as the
explicit config.

This fixes building lxc on NixOS and similar.

This also makes the build more pure and
reproducible. Before this commit building the same
set of checkouts, same config would yield
different lxc artifacts on RedHat and Ubuntu.
It was probably harmless though.

This also removes inactive maintainer from the
Makefile.

Signed-off-by: Michal Kazior <michal@plume.com>